Refactor : refactoring display according to backend data
